refactor(casl): move action.ts
